Moved ProtocolType to types.go and added missing functions
