+++ pkg/network/event_common.go
@@ -14,6 +14,7 @@ import (
 	"github.com/dustin/go-humanize"
 
 	"github.com/DataDog/datadog-agent/pkg/network/dns"
+	"github.com/DataDog/datadog-agent/pkg/network/protocols"
 	"github.com/DataDog/datadog-agent/pkg/network/protocols/http"
 	"github.com/DataDog/datadog-agent/pkg/network/protocols/kafka"
 	"github.com/DataDog/datadog-agent/pkg/process/util"
@@ -266,7 +267,7 @@ type ConnectionStats struct {
 
 	ContainerID *string
 
-	Protocol ProtocolType
+	Protocol protocols.ProtocolType
 }
 
 // Via has info about the routing decision for a flow
